Supervised Learning Algorithms
Applying Supervised Learning algorithms to analyize the following datasets:

DATASET #1
Census Income Data Set. This is a dataset taken from the UCI Repository of a 1994 census data to predict whether the income of an adult exceeds $50k/year based on 14 different attributes such as age, work class, education, occupation, sex, native country, and others. This dataset comes with 48842 different instances which can be used for the classification task.

DATASET #2
Pen-Based Recognition of Handwritten Digits Data Set. This is a dataset taken from the UCI Repository of handwritten digit samples from 44 writers. This is a multiclass dataset with no missing values, 16 attributes and 10992 instances which can be used for the classification task.

Algorithms:
- Decision Trees
- Neural Networks
- Boosting (AdaBoost)
- k-Nearest Neighbors
- Support Vector Machines